+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/Animals/EffectSpawnRandomAnimalFollower.lua
@@ -1,7 +1,7 @@
----@class EffectSpawnRandomAnimal : ChaosEffectBase
-EffectSpawnRandomAnimal = ChaosEffectBase:derive("EffectSpawnRandomAnimal", "spawn_random_animal")
+---@class EffectSpawnRandomAnimalFollower : ChaosEffectBase
+EffectSpawnRandomAnimalFollower = ChaosEffectBase:derive("EffectSpawnRandomAnimalFollower", "spawn_random_animal_follower")
 
-function EffectSpawnRandomAnimal:OnStart()
+function EffectSpawnRandomAnimalFollower:OnStart()
     ChaosEffectBase:OnStart()
 
     local player = getPlayer()
@@ -11,17 +11,14 @@ function EffectSpawnRandomAnimal:OnStart()
     if not square then return end
 
     local animalType, animalBreed = ChaosAnimals.GetRandomAnimal()
-    print("[EffectSpawnRandomAnimal] Spawning animal: " .. animalType .. " " .. animalBreed)
+    print("[EffectSpawnRandomAnimalFollower] Spawning animal: " .. animalType .. " " .. animalBreed)
     local animal = ChaosAnimals.SpawnAnimal(square:getX(), square:getY(), square:getZ(), animalType, animalBreed)
 
     if animal then
         local displayName = getText("IGUI_AnimalType_" .. animal:getAnimalType())
-
         local breedName = getText("IGUI_Breed_" .. animal:getData():getBreed():getName())
-
         local str = string.format(ChaosLocalization.GetString("misc", "spawned_animal"), breedName, displayName)
         ChaosPlayer.SayLine(player, str, 162 / 255, 223 / 255, 208 / 255)
-
         SpecialAnimal:new(animal)
     end
 end